adding toast and message modals + starting developing reviews and booking pages

+++ Pages/User/Cars/Create.cshtml.cs
@@ -113,7 +113,7 @@ namespace VentureCarRentals.Pages.User.Cars
                 return RedirectToPage("/User/Cars/BrowseCars");
             }
 
-            TotalPrice = TotalDays * Car.PricePerDay;
+            TotalPrice = TotalDays * (double)Car.PricePerDay;
 
             return Page();
         }
